+++ CHANGELOG.md
@@ -757,6 +757,10 @@ when an expandable node is selected https://github.com/Textualize/textual/pull/4
 
 ## [0.44.0] - 2023-12-1
 
+### Added
+
+- Added Restrictor class to allow arbitrary text restriction on Input
+
 ### Changed
 
 - Breaking change: Dropped 3.7 support https://github.com/Textualize/textual/pull/3766